rename: eventDetail 파일명 수정
